Check path validity before rename

+++ boilerplate.exs
@@ -59,10 +59,12 @@ defmodule BoilerplateSetup do
       |> Enum.sort_by(&String.length/1)
 
     Enum.each(paths, fn path ->
-      IO.puts("renaming #{path}")
-      new_path = String.replace(path, "boilername", otp_name, global: true)
-      File.mkdir_p!(new_path)
-      File.rename!(path, new_path)
+      if String.contains?("boilername") and File.exists?(path) do
+        IO.puts("renaming #{path}")
+        new_path = String.replace(path, "boilername", otp_name, global: true)
+        File.mkdir_p!(new_path)
+        File.rename!(path, new_path)
+      end
     end)
   end
